backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
psycopg==3.2.3
gunicorn==21.2.0
